refactor(FOROME-725): fixed modal comp req select
