+++ src/pages/filter/page.tsx
@@ -17,8 +17,9 @@ import { ModalNumbers } from '@pages/filter/ui/modal-edit/components/modal-numbe
 import { ErrorPage } from '../error/error'
 import { FilterControl } from './ui/filter-control/filter-control'
 import { ModalCompoundHet } from './ui/modal-edit/components/modal-compound-het/modal-compound-het'
+import { ModalEditCompoundRequest } from './ui/modal-edit/components/modal-compound-request/modal-edit-compound-request'
+import { ModalSelectCompoundRequest } from './ui/modal-edit/components/modal-compound-request/modal-select-compound-request'
 import { ModalCustomInheritanceMode } from './ui/modal-edit/components/modal-custom-inheritance-mode/modal-custom-inheritance-mode'
-import { ModalEditCompoundRequest } from './ui/modal-edit/components/modal-edit-compound-request'
 import { ModalEditFilters } from './ui/modal-edit/components/modal-edit-filters'
 import { ModalEditGeneRegion } from './ui/modal-edit/components/modal-edit-gene-region'
 import { ModalEditInheritanceMode } from './ui/modal-edit/components/modal-inheritance-mode/modal-edit-inheritance-mode'
@@ -27,7 +28,6 @@ import { ModalTextEditor } from './ui/query-builder/modal-text-editor'
 import { QueryBuilder } from './ui/query-builder/query-builder'
 import { ModalSaveDataset } from './ui/query-builder/ui/modal-save-dataset'
 import { ModalSelectAttribute } from './ui/query-builder/ui/modal-select-attribute'
-import { ModalSelectCompoundRequest } from './ui/query-builder/ui/modal-select-compound-request'
 import { ModalSelectFilters } from './ui/query-builder/ui/modal-select-filters'
 import { ModalSelectGeneRegion } from './ui/query-builder/ui/modal-select-gene-region'
 import { getNumberWithCommas } from './ui/query-builder/ui/next-step-route'
