+++ src/pages/filter/ui/modal-edit/components/modal-compound-het/modal-compound-het.store.ts
@@ -1,10 +1,9 @@
-import { makeAutoObservable, toJS } from 'mobx'
+import { makeAutoObservable } from 'mobx'
 
 import { ActionType } from '@declarations'
 import dtreeStore from '@store/dtree'
 import { addAttributeToStep } from '@utils/addAttributeToStep'
 import { changeFunctionalStep } from '@utils/changeAttribute/changeFunctionalStep'
-import { getQueryBuilder } from '@utils/getQueryBuilder'
 import modalEditStore, { IParams } from '../../modal-edit.store'
 
 class ModalCompoundHetStore {
@@ -27,44 +26,6 @@ class ModalCompoundHetStore {
   public resetConditions(): void {
     this.stateCondition = '-current-'
     this.approxCondition = 'transcript'
-  }
-
-  public get attrData() {
-    let attrData: any
-
-    const subGroups = Object.values(
-      getQueryBuilder(toJS(dtreeStore.startDtreeStat['stat-list'])),
-    )
-
-    subGroups.map(subGroup => {
-      subGroup.map((item, currNo) => {
-        if (item.name === modalEditStore.groupName) {
-          attrData = subGroup[currNo]
-        }
-      })
-    })
-
-    return attrData
-  }
-
-  public get approxOptions(): string[] {
-    const approxOptions: string[] = []
-
-    this.attrData['approx-modes'].map((mode: string[]) => {
-      approxOptions.push(mode[1])
-    })
-
-    return approxOptions
-  }
-
-  public get approxValues(): string[] {
-    const approxValues: string[] = []
-
-    this.attrData['approx-modes'].map((mode: string[]) => {
-      approxValues.push(mode[0])
-    })
-
-    return approxValues
   }
 
   public openModalAttribute(): void {
