+++ src/pages/filter/ui/modal-edit/components/modal-compound-request/modal-edit-compound-request.tsx
@@ -10,20 +10,20 @@ import { Button } from '@ui/button'
 import { InputNumber } from '@ui/input-number'
 import { Select } from '@ui/select'
 import { DecisionTreeModalDataCy } from '@components/data-testid/decision-tree-modal.cy'
+import { resetOptions } from '@pages/filter/ui/panels/function-panel/components/compound-request/approx-state'
+import { AllNotModalMods } from '@pages/filter/ui/query-builder/ui/all-not-modal-mods'
+import { ApproxStateModalMods } from '@pages/filter/ui/query-builder/ui/approx-state-modal-mods'
+import { DisabledVariantsAmount } from '@pages/filter/ui/query-builder/ui/disabled-variants-amount'
+import { HeaderModal } from '@pages/filter/ui/query-builder/ui/header-modal'
+import { ModalBase } from '@pages/filter/ui/query-builder/ui/modal-base'
 import { changeFunctionalStep } from '@utils/changeAttribute/changeFunctionalStep'
 import { getFuncParams } from '@utils/getFuncParams'
 import { getRequestData } from '@utils/getRequestData'
 import { getResetRequestData } from '@utils/getResetRequestData'
 import { getResetType } from '@utils/getResetType'
 import { getSortedArray } from '@utils/getSortedArray'
-import { resetOptions } from '../../panels/function-panel/components/compound-request/compound-request'
-import { AllNotModalMods } from '../../query-builder/ui/all-not-modal-mods'
-import { ApproxStateModalMods } from '../../query-builder/ui/approx-state-modal-mods'
-import { DisabledVariantsAmount } from '../../query-builder/ui/disabled-variants-amount'
-import { HeaderModal } from '../../query-builder/ui/header-modal'
-import { ModalBase } from '../../query-builder/ui/modal-base'
-import { IParams, selectOptions } from '../modal-edit.store'
-import { EditModalButtons } from './edit-modal-buttons'
+import { IParams, selectOptions } from '../../modal-edit.store'
+import { EditModalButtons } from '../edit-modal-buttons'
 
 export const ModalEditCompoundRequest = observer((): ReactElement => {
   const ref = useRef(null)
